--- tests/unit/test_charm.py
@@ -1,13 +1,0 @@
-# Copyright 2024 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Unit tests for the charm."""
-
-
-def test_nothing():
-    """
-    arrange: nothing.
-    act: nothing.
-    assert: True.
-    """
-    assert True
